+++ apps/web/pages/community/[id].tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from 'react'
 
 import { CommunityDetailHeader } from '@/components/domain/community'
-import { Layout } from '@/components/public/layout'
+import Layout from '@/components/public/Layout'
 import CommunityDetailScreen from '@/components/screens/ScreenCommunityDetail'
 
 export default function CommunityDetailPage() {
