+++ apps/web/pages/community/generate.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from 'react'
 
 import { CommunityGenerateHeader } from '@/components/domain/community/generate/header'
-import Layout from '@/components/public/layout/Layout'
+import Layout from '@/components/public/Layout'
 import CommunityPostGenerateScreen from '@/components/screens/ScreenCommunityPostGenerate'
 
 export default function Generate() {
