+++ apps/web/pages/community/index.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from 'react'
 
 import { CommunityHeader } from '@/components'
-import { Layout } from '@/components/public/layout'
+import Layout from '@/components/public/Layout'
 import CommunityMainScreen from '@/components/screens/MainCommunity'
 
 export default function CommunityPage() {
