+++ apps/web/pages/complaint/[id].tsx
@@ -1,7 +1,7 @@
 import { type ReactElement } from 'react'
 
 import { ComplaintGenerateHeader } from '@/components/domain/complaint'
-import { Layout } from '@/components/public/layout'
+import Layout from '@/components/public/Layout'
 import ComplaintDetailScreen from '@/components/screens/ScreenComplaintDetail'
 
 export default function LostDetailPage() {
