+++ apps/web/pages/lost/generate.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from 'react'
 
 import { LostGenerateHeader } from '@/components'
-import Layout from '@/components/public/layout/Layout'
+import Layout from '@/components/public/Layout'
 import LostPostGenerateScreen from '@/components/screens/ScreenLostPostGenerate'
 
 export default function Generate() {
