+++ apps/web/src/components/domain/alarm/header/AlarmHeader.tsx
@@ -1,4 +1,4 @@
-import { Header } from '@/components/public/layout'
+import { Header } from '@/components/public/header'
 
 export const AlarmHeader = () => {
   return (
