+++ apps/web/src/components/domain/alarm/index.ts
@@ -1,1 +1,3 @@
 export * from './header'
+export { default as DirectMessages } from './DirectMessages'
+export { default as Notifications } from './Notifications'
