+++ apps/web/src/components/domain/community/generate/header/CommunityGenerateHeader.tsx
@@ -1,6 +1,6 @@
 import { useBoolean } from '@ahhachul/lib'
 import { Dialog } from '@/components/common/dialog'
-import { Header } from '@/components/public/layout'
+import { Header } from '@/components/public/header'
 
 export const CommunityGenerateHeader = () => {
   const [isDialogShowing, _, onDialogOpen, onDialogClose] = useBoolean(false)
