+++ apps/web/src/components/domain/setting/nickname/header/NicknameHeader.tsx
@@ -1,4 +1,4 @@
-import { Header } from '@/components/public/layout'
+import { Header } from '@/components/public/header'
 
 const SettingNicknameHeader = () => {
   return <Header title="로그인" hasGoBack goBackToHome />
