+++ apps/web/src/components/public/MonitoringInitializer.tsx
