+++ apps/web/src/components/public/cta/BottomSheetForLogin.tsx
@@ -1,13 +1,42 @@
 import { css } from '@emotion/react'
 import styled from '@emotion/styled'
+import Image from 'next/image'
+import LogoImg from 'public/illust/img/img_favicon.png'
+import { forwardRef } from 'react'
 
 import { BottomSheet } from '@/components/common'
+import { PATH } from '@/constants'
 
-export const CTABottomSheet = styled(BottomSheet)`
+interface BottomSheetForLoginProps {
+  isOpen: boolean
+  onClose: () => void
+}
+
+const BottomSheetForLogin = forwardRef<HTMLDialogElement, BottomSheetForLoginProps>(({ isOpen, onClose }, ref) => {
+  return (
+    <CTABottomSheet ref={ref} isHeaderHidden title="앱 다운로드 팝업" isOpen={isOpen} onClose={onClose}>
+      <ContentBox>
+        <Image src={LogoImg} alt="" width={94} height={94} />
+        <Strong>
+          아하철에서 제공되는 혜택을
+          <br />
+          놓치고 계신 건 아닌가요?
+        </Strong>
+        <P>아하철 로그인시 다양한 혜택 및 팁을 제공 중!</P>
+        <Link href={PATH.LOGIN}>로그인하기</Link>
+        <CloseBtn type="button" onClick={onClose}>
+          괜찮아요, 다음에 로그인 할게요.
+        </CloseBtn>
+      </ContentBox>
+    </CTABottomSheet>
+  )
+})
+
+const CTABottomSheet = styled(BottomSheet)`
   padding: 70px 20px 46px;
 `
 
-export const ContentBox = styled.div`
+const ContentBox = styled.div`
   width: 100%;
   text-align: center;
   & > img {
@@ -15,7 +44,7 @@ export const ContentBox = styled.div`
   }
 `
 
-export const Strong = styled.strong`
+const Strong = styled.strong`
   ${({ theme }) => css`
     ${theme.fonts.bold20};
     display: block;
@@ -25,7 +54,7 @@ export const Strong = styled.strong`
   `}
 `
 
-export const P = styled.p`
+const P = styled.p`
   ${({ theme }) => css`
     ${theme.fonts.regular14};
     padding: 8px 0 24px;
@@ -33,7 +62,7 @@ export const P = styled.p`
   `}
 `
 
-export const Link = styled.a`
+const Link = styled.a`
   ${({ theme }) => css`
     ${theme.fonts.semibold14};
     display: flex;
@@ -65,7 +94,7 @@ export const Link = styled.a`
   `}
 `
 
-export const CloseBtn = styled.button`
+const CloseBtn = styled.button`
   ${({ theme }) => css`
     ${theme.fonts.medium14};
     width: 100%;
@@ -76,3 +105,5 @@ export const CloseBtn = styled.button`
     background-color: ${theme.colors.white};
   `}
 `
+
+export default BottomSheetForLogin
