+++ apps/web/src/components/public/gnb/index.ts
@@ -1,0 +1,1 @@
+export { default as NavBar } from './Navbar'
