+++ apps/web/src/components/public/index.ts
@@ -1,1 +1,2 @@
+export { default as AppInner } from './AppInner'
 export { default as MonitoringInitializer } from './MonitoringInitializer'
