--- apps/web/src/components/public/layout/index.ts
@@ -1,2 +1,0 @@
-export * from './header'
-export { default as Layout } from './Layout'
